feat: boot the kernel as a wasm32 instance
The arch-independent kernel now runs as a single WebAssembly instance in
Node,
printing its banner through a host import and panicking correctly. First
milestone of the arch-wasm32 port (ARCH_WASM32.md §11; M0 landed
earlier).

**New crates**
- `arch-wasm32`: the HAL, with console, clock, and the halt/exit path as
  `env`
  imports. Everything with no host boundary — paging stubs, frame
  layout, VA
  layout, the page arena — is re-exported from `arch-sim` rather than
  duplicated
  a fourth time, which is how the design frames the two: the simulator
  *is* this
  port minus the import boundary.
- `kernel-wasm`: the platform layer, `kernel-boot`'s counterpart at the
  host
  boundary. Owns the two entry points the host calls and the panic
  handler.
  Excluded from the workspace, since it is a wasm-only binary.

**Kernel changes for a 32-bit pointer target**
- `vm.rs`: `NR_PHYS_PAGES` computed through `u64`, because the 32 GiB
  literal
  does not fit a 32-bit `usize`.
- `ipc.rs`: the senda size field is read as `u64` and narrowed, since
  `usize::from_le_bytes` wants four bytes there and stopped compiling.
- `clock.rs`: `MinixTimer`'s layout assertions apply only where a
  pointer is
  eight bytes, rather than asserting a size the struct cannot have on
  wasm32.
- `syscall.rs`: the `asm!` compiler barrier becomes a `compiler_fence`
  on
  wasm32, where inline assembly is unstable.
- `consts.rs`: a wasm32 arm for the sigframe offsets, which existed only
  for the
  three hardware arches; `arch-sim`'s frame builder now uses those
  offsets.

  All five are ABI-neutral on the 64-bit arches and the host kernel
  suite still
  passes after them. The general lesson is recorded in §5.2 as an
  invariant for
  M2: message and ABI structs must carry `u64` handles and fixed-width
  integers,
  never pointers or `usize` — a struct that violates this agrees with
  itself on
  any one target and silently disagrees with a server compiled for
  another.

**Harness (`tools/wasm-boot/`)**
- Instantiates the module with a simulated host and checks both phases:
  the
  banner arriving through `host_console_write`, and a deliberate panic
  reporting
  its message and source location before halting and trapping.
- Records that the import boundary is the *used* set, not the declared
  one:
  `--gc-sections` drops the two console-input imports because M1 has no
  input
  path, so the import list is not a contract.

Verified: `sh tools/wasm-boot/run.sh` → 4/4 checks passed;
`kernel-sim-tests`
still 24/24; clippy clean on `arch-wasm32`, `arch-sim`, `arch-common`,
`kernel --features sim`, and `kernel-wasm` for wasm32. The three QEMU
arches were
not rebuilt, which needs the stage1 fork toolchain.

+++ crates/kernel/src/ipc.rs
@@ -1297,7 +1297,10 @@ pub unsafe fn ipc_notify_handler(caller: *mut Proc, msg: &mut [u8; MESSAGE_SIZE]
 pub unsafe fn ipc_senda_handler(caller: *mut Proc, msg: &mut [u8; MESSAGE_SIZE]) -> i32 {
     unsafe {
         let table = u64::from_le_bytes(msg[8..16].try_into().unwrap_or([0; 8])) as *mut u8;
-        let size = usize::from_le_bytes(msg[16..24].try_into().unwrap_or([0; 8]));
+        // The field holds eight bytes in the message, so read it as u64 and
+        // narrow. `usize::from_le_bytes` would demand four bytes on a 32-bit
+        // target and stop compiling.
+        let size = u64::from_le_bytes(msg[16..24].try_into().unwrap_or([0; 8])) as usize;
         try_deliver_senda(caller, table, size)
     }
 }
